refactor folders and remove deprecated
