--- config/config.txt
@@ -1,46 +1,0 @@
-[CONTA]
-email = exemplo@email.com
-senha = senha
-tipo_conta = Treino
-
-[ENTRADAS]
-arquivo = entradas.txt
-tipo_par = Auto
-valor = 100
-tempo = 1
-profit_minimo = 0
-
-[WIN]
-stopwin = 300
-max_soros = 2
-scalper_win = 0
-
-[LOSS]
-stoploss = 1000
-max_gale = 2
-tipo_gale = martin
-tipo_martin = seguro
-vez_gale = sinal
-scalper_loss = 0
-
-[AJUSTES]
-correcao_entrada = 2
-delay = False
-
-[TENDENCIA]
-tendencia = False
-tipo_tendencia = simples
-periodo_tendencia = 21
-
-[NOTICIAS]
-toros = 0
-noticias_hora = 10
-noticias_minuto = 0
-
-[ESTRATEGIA]
-auto = False
-autotime = 1
-autogale = 2
-estrategia = MHI
-tipo_milhao = Minoria
-paridade = EURUSD
