--- config/entradas.txt
@@ -1,10 +1,0 @@
-EURUSD PUT 21:40
-EURUSD CALL 21:41
-EURUSD PUT 21:42
-EURUSD CALL 21:43
-EURUSD PUT 21:44
-EURUSD CALL 21:45
-EURUSD PUT 21:46
-EURUSD CALL 21:47
-EURUSD PUT 21:48
-EURUSD CALL 21:49
