--- misc/ajuda.txt
@@ -1,90 +1,0 @@
-Versão 6 - API updated
-
-            [RUN]
-Para rodar
-        python bot.py
-Caso queira saber os demais comandos do bot:
-        python bot.py -h
-
-            [CONFIG]
-Caso quiser configurar via interface gráfica, rode o arquivo settings.py
-Arquivo de configuração:
- - As opções estarão na forma de primeira|segunda, escolha uma ou outra.
- - Caso tiver valor-outro significa uma sequência de qualquer valor. E inf significa infinito.
-O que significa:
- - email: email da conta, ou número
- - senha: senha da conta
- - tipo_conta: se é a conta treino ou real
- - arquivo: nome do arquivo que vai estar as entradas
- - tipo_par: se a paridade é binaria/digital, se colocar auto, ele vai ver qual está disponível, ou tem maior profit
- - valor: o valor das entradas
- - tempo: timeframe padrão das entradas
- - minimo: payout minimo para fazer a entrada
- - correcao: segundos para se descontar do tempo de entrada
- - stopwin: valor objetivo que para se alcançar
- - stoploss: valor limite de perdas
- - max_gale: máximo de martingales
- - tipo_gale: caso haver gale, rode o script com -m
- - delay: tempo para pegar o resultado
- - vez_gale: se o martingale irá ser na próxima vela ou entrada
- - tendencia: se vai querer negar entradas caso estiver contra a tendencia
- - tipo_tendencia: se vai verificar as últimas 3 velas (velas) ou não (simples)
- - periodo_tendencia: quantas velas serão analisadas na tendência para o SMA
- - toros: filtrar por número de estrelas nas notícias
- - noticias_hora: se tiver alguma notícia x horas antes/depois não irá entrar
- - noticias_minuto: se tiver alguma notícia x minutos antes/depois não irá entrar
- - auto: ativar o autotrade
- - autotime: timeframe do trade automático
- - autogale: procurar estratégias com no máximo 0-2 gales de acerto
- - estrategia: a estratégia caso não estiver automático
- - tipo_milhao: em estratégias pertinentes, se vai operar pela maioria/minoria
- - paridade: a paridade da estratégia
-
-Exemplo:
-
-[CONTA]
-email = exemplo@email.com
-senha = exemplo
-tipo_conta = treino|real
-
-[ENTRADAS]
-arquivo = entradas.txt
-tipo_par = binaria|digital|auto
-valor = 1-inf
-tempo = 1|5|15
-profit_minimo = 0-100
-
-[WIN]
-stopwin = 2-inf
-max_soros = 0-inf
-scalper_win = 0-100
-
-[LOSS]
-stoploss = 1-inf
-martin = True|False
-max_gale = 1-inf
-scalper_loss = 0-100
-tipo_gale = seguro|simples|leve|agressivo|2.0-inf
-vez_gale = vela|entrada
-
-[AJUSTES]
-correcao_entrada = 0-inf
-delay = False|-inf-inf
-
-[TENDENCIA]
-tendencia = False|True
-tipo_tendencia = sma|velas
-periodo_tendencia = 1-inf
-
-[NOTICIAS]
-toros = 0-3
-noticias_hora = 0-inf
-noticias_minuto = 0-inf
-
-[ESTRATEGIA]
-auto = True|False
-autotime = 1|5|15
-autogale = 0-2
-estrategia = -
-tipo_milhao = Maioria|Minoria
-paridade = -
