+++ requirements.txt
@@ -1,19 +1,9 @@
-aiohttp==3.6.2
-amanobot==2.0.1
-async-timeout==3.0.1
-attrs==19.3.0
-beautifulsoup4==4.9.1
+amanobot==2.1.0
+beautifulsoup4==4.9.3
 bs4==0.0.1
-certifi==2020.6.20
-chardet==3.0.4
-dnspython==1.16.0
-idna==2.10
-multidict==4.7.6
+dnspython==2.1.0
 numpy==1.19.1
-pymongo==3.10.1
+pymongo==3.11.4
 requests==2.24.0
-six==1.15.0
-soupsieve==2.0.1
 urllib3==1.25.9
 websocket-client==0.56.0
-yarl==1.4.2
