--- schema/waiting_schema.py
@@ -1,1 +1,0 @@
-queue = { 'email': '' }
